add support for inline queries
This commit adds support for receiving `InlineQuery` messages from
Telegram and let us respond to them with `answerInlineQuery`. There are
however some limitations.

We do not add support for `ChosenInlineResult` at this point, so there
also no classes for all the different types of `InlineResult`. This
means we do not support turning on "inline feedback" via BotFather. We
also don't support WebApps, so all the functionality relating to this
has not been implemented for inline queries.